chore: address bugbot comments
