+++ src/features/dashboard/sandbox/inspect/dir.tsx
@@ -4,7 +4,7 @@ import { cn } from '@/lib/utils'
 import { DataTableRow } from '@/ui/data-table'
 import { AlertCircle, FolderClosed, FolderOpen } from 'lucide-react'
 import { AnimatePresence, motion } from 'motion/react'
-import SandboxInspectEmptyNode from './empty'
+import SandboxInspectEmptyNode from './empty-node'
 import { FilesystemNode } from './filesystem/types'
 import { useDirectory } from './hooks/use-directory'
 import SandboxInspectNode from './node'
